gpuav: Remoe InjectConditionalFunctionPass
