+++ BUILD.gn
@@ -223,8 +223,6 @@ vvl_sources = [
   "layers/gpuav/spirv/link.h",
   "layers/gpuav/spirv/module.cpp",
   "layers/gpuav/spirv/module.h",
-  "layers/gpuav/spirv/inject_conditional_function_pass.cpp",
-  "layers/gpuav/spirv/inject_conditional_function_pass.h",
   "layers/gpuav/spirv/pass.cpp",
   "layers/gpuav/spirv/pass.h",
   "layers/gpuav/spirv/ray_query_pass.cpp",
